Implemented Leaderboard Backend

+++ backend/src/main/java/com/shareapp/DataTransferObjects/LeaderboardDTO.java
@@ -1,11 +1,12 @@
 package com.shareapp.DataTransferObjects;
+import java.math.BigDecimal;
 
 public class LeaderboardDTO {
     String userName;
-    Double totalValue;
+    BigDecimal totalValue;
     int rank;
 
-    public LeaderboardDTO(String username, Double totalValue, int rank){
+    public LeaderboardDTO(String username, BigDecimal totalValue, int rank){
         this.userName = username;
         this.totalValue = totalValue;
         this.rank = rank;
